requirements.txt
streamlit
google-generativeai
python-dotenv
lanchain 
langchain_google_genai